Fix errors in the Parkinglot state machine and its implementation
The pay event lead to the "paused-credit=1"-state instead of the correct
"paused-credit>1"-state when in the "active-credit=1"-state.
Furthermode there was a bug in the special transition code ("credit>1" to
"credit=1" and the latter to "credit=0"), which was executed for all events
and not just for tick events. The second special case was also not necessary
and was thus removed.